requirements.txt
asgiref==3.2.10
Django==3.1.2
Pillow==7.2.0
python-barcode==0.13.1
pytz==2020.1
sqlparse==0.4.1
